+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1-beta-5] - unreleased
+- added initial Violin plot support
+- improbed boxplot support
+
 ## [1-beta-4] - 2024-12-14
 - improved styling of `imshow`
 - added initial 3d `surface` support